Finished third activity
Its more on the array methods that did the job. first string to array by
using split and using no space as parameter which basically takes every
letter and space as an individual elements. Then use the reverse method
for the array then use join method to with parameter no space to convert
it back to string