Finished 4th activity with the TOP solution
Still not used using this methods. so basically create a new array for
container then use foreach method to loop each item in the array then
in the function expression  used if to determine if the element in the
args array includes the item if it does not push method the element to
the container array earlier after that return the newArray.