Finished 6th activity
I did not got the right answer which is disappointing but understood
on how this function works. first to compare if divisible by 4 then
use && to consider the other part even if its true in the first part.
Then use || to either if true. this is where I got the part I didnt got
use !== to for the divisible by 100 part cause if the year aint divi by
100 it return true  even it says false it will compare again to divi by
400 so logically it is divi by 100 and 400 thus a leap year.